kvnemesis: rename from kvnemeses
It was originally named "kvnemeses" (plural, more than one nemesis) to
be parallel to the cdc nemeses tests, which it was modeled after. But
I've found others (and myself) calling it "kvnemesis" and it reads much
better, so rename it before we get too much git history at the old
package name.

Release note: None
